requirements.txt
discord.py==2.3.2
jishaku==2.5.2
requests==2.31.0